+++ back/src/main/java/com/back/domain/delivery/delivery/repository/DeliveryRepository.java
@@ -3,5 +3,5 @@ package com.back.domain.delivery.delivery.repository;
 import com.back.domain.delivery.delivery.entity.Delivery;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DeliveryRepository extends JpaRepository<Delivery, Integer> {
+public interface DeliveryRepository extends JpaRepository<Delivery, Long> {
 }
